// backend/src/common/errors/deployment-error-code.enum.ts
export enum DeploymentErrorCode {
  DEP_RELEASE_INVALID = 'DEP_RELEASE_INVALID',
  DEP_RELEASE_NOT_FOUND = 'DEP_RELEASE_NOT_FOUND',
  DEP_RELEASE_EXISTS = 'DEP_RELEASE_EXISTS',
  DEP_IMMUTABLE_RELEASE = 'DEP_IMMUTABLE_RELEASE',
  DEP_RELEASE_LIFECYCLE_INVALID = 'DEP_RELEASE_LIFECYCLE_INVALID',

  DEP_GATE_FAILED = 'DEP_GATE_FAILED',
  DEP_GATE_NOT_FOUND = 'DEP_GATE_NOT_FOUND',
  DEP_GATE_BYPASS_FORBIDDEN = 'DEP_GATE_BYPASS_FORBIDDEN',

  DEP_ARTIFACT_MISSING = 'DEP_ARTIFACT_MISSING',
  DEP_CONTRACT_INCOMPATIBLE = 'DEP_CONTRACT_INCOMPATIBLE',

  DEP_DEPLOYMENT_FAILED = 'DEP_DEPLOYMENT_FAILED',
  DEP_DEPLOYMENT_NOT_FOUND = 'DEP_DEPLOYMENT_NOT_FOUND',
  DEP_DEPLOYMENT_CANCELLED = 'DEP_DEPLOYMENT_CANCELLED',
  DEP_DEPLOYMENT_CANNOT_CANCEL = 'DEP_DEPLOYMENT_CANNOT_CANCEL',

  DEP_HEALTHCHECK_FAILED = 'DEP_HEALTHCHECK_FAILED',
  DEP_ROLLBACK_FAILED = 'DEP_ROLLBACK_FAILED',
  DEP_ROLLBACK_NO_TARGET = 'DEP_ROLLBACK_NO_TARGET',

  DEP_PERMISSION_DENIED = 'DEP_PERMISSION_DENIED',
  DEP_CONCURRENCY_LOCKED = 'DEP_CONCURRENCY_LOCKED',
  DEP_ENVIRONMENT_LOCKED = 'DEP_ENVIRONMENT_LOCKED',
  DEP_ENVIRONMENT_NOT_FOUND = 'DEP_ENVIRONMENT_NOT_FOUND',
  DEP_PROMOTION_INVALID_STAGE = 'DEP_PROMOTION_INVALID_STAGE',
  DEP_INTERNAL_ERROR = 'DEP_INTERNAL_ERROR',
}